pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "music-downloader"
version = "0.1.0"
description = "Automated music discovery and download via Telegram bot. Resolves metadata from Spotify, searches and downloads FLAC from Soulseek (slskd), and organizes your library."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "Sergio Fernández", email = "sergio@geiser.cloud"}
]
keywords = ["music", "downloader", "telegram", "soulseek", "slskd", "flac", "spotify"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    "python-telegram-bot>=22.6",
    "spotipy>=2.25.2",
    "slskd-api>=0.1.5",
    "mutagen>=1.47.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "ruff>=0.9.0",
    "pre-commit>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/GeiserX/music-downloader"
Repository = "https://github.com/GeiserX/music-downloader"
Issues = "https://github.com/GeiserX/music-downloader/issues"
Changelog = "https://github.com/GeiserX/music-downloader/blob/master/docs/CHANGELOG.md"

[project.scripts]
music-downloader = "music_downloader.__main__:main"

[tool.setuptools]
packages = ["music_downloader"]
package-dir = {"music_downloader" = "src/music_downloader"}

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (FastAPI depends on this)
    "SIM102", # collapsible-if (readability preference)
    "SIM108", # use ternary operator (readability preference)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F841"]

[tool.ruff.lint.isort]
known-first-party = ["music_downloader"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
